+++ tests/xharness/Xharness.Tests/Xharness.Tests.csproj
@@ -61,6 +61,8 @@
     <Compile Include="TestImporter\Xamarin\Tests\ProjectFilterTest.cs" />
     <Compile Include="TestImporter\Xamarin\Tests\AssemblyLocatorTest.cs" />
     <Compile Include="TestImporter\Xamarin\Tests\XamariniOSTemplateTest.cs" />
+    <Compile Include="Tests\TestTargetExtensionsTests.cs" />
+    <Compile Include="Tests\TestPlatformExtensionsTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
