+++ tests/xharness/xharness.csproj
@@ -125,6 +125,10 @@
     <Compile Include="TestTasks\IRunSimulatorTask.cs" />
     <Compile Include="TestTasks\RunDevice.cs" />
     <Compile Include="TestProjectExtensions.cs" />
+    <Compile Include="Jenkins\TestVariationsFactory.cs" />
+    <Compile Include="Jenkins\TestData.cs" />
+    <Compile Include="TestTargetExtensions.cs" />
+    <Compile Include="TestPlatformExtensions.cs" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="..\..\tools\common\SdkVersions.cs">
